AI-Doc-Manager/backend_feature_chat/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0                  # async PostgreSQL driver (Cloud SQL)
cloud-sql-python-connector[asyncpg]==1.9.0  # Google Cloud SQL Connector

# ── Google Cloud ──────────────────────────────────────────────────────────────
google-auth==2.29.0
google-cloud-aiplatform==1.52.0  # Vertex AI SDK (Gemini + Vector Search)
google-cloud-storage==2.16.0     # GCS
firebase-admin==6.5.0            # Verify Firebase ID Token

# ── AI / ADK ──────────────────────────────────────────────────────────────────
google-adk==0.3.0                # Google Agent Development Kit
vertexai==1.52.0                 # Vertex AI convenience wrapper

# ── Validation & config ───────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx==0.27.0
tenacity==8.3.0                  # retry logic cho Vertex AI calls
structlog==24.1.0                # structured logging cho Cloud Logging
